Move excerpt maker into string category
